Cleanup.

+++ src/command/verify/protocol.c
@@ -43,8 +43,7 @@ verifyProtocol(const String *command, const VariantList *paramList, ProtocolServ
                 varUInt64(varLstGet(paramList, 2)),                                                 // File size
                 varStr(varLstGet(paramList, 3)));                                                   // Cipher pass
 
-            // !!! HIGHLY SUSPECT -- A STRUCT IS BEING RETURNED BUT INT IS PASSED BACK
-            protocolServerResponse(server, varNewInt((int)result));
+            protocolServerResponse(server, varNewInt(result));
         }
         else
             found = false;
